requirements.txt
fastapi==0.100.0
hypercorn==0.14.4
fastapi[all]
sqlalchemy
asyncpg
databases
psycopg2
python-dotenv
pymongo
pyjwt 
passlib[bcrypt]